Extract workspace-status helpers into modules/workspace_status.py + modules/dependency_reasons.py
This is PR A from the quickstart.py refactor audit -- the keystone move that
unlocks downstream blueprint extractions. The 21 workspace-status functions and
26 dependency-reason functions were inseparable in practice (the workspace
builder calls every dependency-reason adapter), so they're extracted as two
sibling modules.

Why split into two modules instead of one:
  * Each function family is large enough on its own (679 + 325 functional lines)
    that keeping them together would push past ~1,200 lines.
  * The split matches the natural import dependency: workspace_status imports
    from dependency_reasons, not vice versa.
  * dependency_reasons is purely libraries-data inspection (no Flask, no DB, no
    helpers) and will be reused by the future library_routes blueprint.

Three tiny helpers (_normalize_status, _is_truthy_setting_value,
_parse_json_array) moved to dependency_reasons.py since 95% of their callers
live there. They are re-exported via quickstart.py for the handful of remaining
in-file callers.

Constants moved alongside their helpers:
  * dependency_reasons.py owns QS_*_DEP_*, QS_*_OVERLAY_IMAGE_VALUES,
    QS_*_REQUIRED_STEP_KEY (the per-service step keys).
  * workspace_status.py owns QS_REQUIRED_STEP_KEYS, QS_REVIEW_STEP_KEYS,
    QS_VALIDATION_STEP_KEYS, QS_STATUS_ORDER, QS_WARN_REASONS, QS_ERROR_REASONS,
    QS_FINAL_VALIDATION_TTL_HOURS.

quickstart.py re-exports every public name so external tests using
`qs_module._build_workspace_status_context` etc. keep working. The change
does, however, require tests that mock workspace_status-internal callees
(e.g. _build_final_gate, _build_kometa_install_context,
_get_imagemaid_settings_section) to patch modules.workspace_status directly,
because those callees are resolved through the workspace_status module's own
namespace -- not through quickstart's. A new workspace_status_module fixture in
tests/conftest.py makes that explicit.

Size impact:
  * quickstart.py: 11,087 -> 9,831 lines (-1,256 lines / -11.3%)
  * modules/workspace_status.py: 879 lines (over the 600 soft-limit but
    cohesive; further split would harm cohesion)
  * modules/dependency_reasons.py: 593 lines

Tests: 791 passing (zero regressions). Updated 4 tests in
test_workspace_dependency_logic.py to patch the workspace_status module instead
of qs_module for internal references.
