+++ tests/test_workspace_dependency_logic.py
@@ -523,12 +523,12 @@ def test_workspace_status_route_returns_all_dependency_reasons(client, monkeypat
     assert "120-sonarr" in payload["required_keys"]
 
 
-def test_workspace_app_readiness_kometa_points_to_first_blocker(monkeypatch, qs_module):
+def test_workspace_app_readiness_kometa_points_to_first_blocker(monkeypatch, qs_module, workspace_status_module):
     monkeypatch.setattr(qs_module.helpers, "get_menu_list", _template_list)
     monkeypatch.setattr(qs_module.database, "get_unique_config_names", lambda: ["cfg"])
-    monkeypatch.setattr(qs_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
+    monkeypatch.setattr(workspace_status_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
     monkeypatch.setattr(
-        qs_module,
+        workspace_status_module,
         "_build_final_gate",
         lambda *_args, **_kwargs: {
             "stage": "todo",
@@ -536,12 +536,12 @@ def test_workspace_app_readiness_kometa_points_to_first_blocker(monkeypatch, qs_
             "todo_blockers": [{"key": "020-tmdb", "label": "TMDb"}],
         },
     )
-    monkeypatch.setattr(qs_module, "_build_kometa_install_context", lambda _name: {})
-    monkeypatch.setattr(qs_module, "_get_imagemaid_settings_section", lambda _name: ({}, {}))
+    monkeypatch.setattr(workspace_status_module, "_build_kometa_install_context", lambda _name: {})
+    monkeypatch.setattr(workspace_status_module, "_get_imagemaid_settings_section", lambda _name: ({}, {}))
     monkeypatch.setattr(qs_module.helpers, "get_imagemaid_root_path", lambda: "")
-    monkeypatch.setattr(qs_module, "_probe_imagemaid_root_state", lambda _path: {})
+    monkeypatch.setattr(workspace_status_module, "_probe_imagemaid_root_state", lambda _path: {})
     monkeypatch.setattr(qs_module.database, "retrieve_section_data", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr(qs_module, "_validate_imagemaid_settings", lambda *_args, **_kwargs: (False, "", ""))
+    monkeypatch.setattr(workspace_status_module, "_validate_imagemaid_settings", lambda *_args, **_kwargs: (False, "", ""))
 
     payload = qs_module._build_workspace_app_readiness("cfg")
 
@@ -551,12 +551,12 @@ def test_workspace_app_readiness_kometa_points_to_first_blocker(monkeypatch, qs_
     assert payload["kometa"]["target_step"] == "020-tmdb"
 
 
-def test_workspace_app_readiness_imagemaid_missing_plex_points_to_plex(monkeypatch, qs_module):
+def test_workspace_app_readiness_imagemaid_missing_plex_points_to_plex(monkeypatch, qs_module, workspace_status_module):
     monkeypatch.setattr(qs_module.helpers, "get_menu_list", _template_list)
     monkeypatch.setattr(qs_module.database, "get_unique_config_names", lambda: ["cfg"])
-    monkeypatch.setattr(qs_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
+    monkeypatch.setattr(workspace_status_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
     monkeypatch.setattr(
-        qs_module,
+        workspace_status_module,
         "_build_final_gate",
         lambda *_args, **_kwargs: {
             "stage": "config",
@@ -564,17 +564,17 @@ def test_workspace_app_readiness_imagemaid_missing_plex_points_to_plex(monkeypat
             "todo_blockers": [],
         },
     )
-    monkeypatch.setattr(qs_module, "_build_kometa_install_context", lambda _name: {})
-    monkeypatch.setattr(qs_module, "_get_imagemaid_settings_section", lambda _name: ({"validated": False}, {"imagemaid": {}}))
+    monkeypatch.setattr(workspace_status_module, "_build_kometa_install_context", lambda _name: {})
+    monkeypatch.setattr(workspace_status_module, "_get_imagemaid_settings_section", lambda _name: ({"validated": False}, {"imagemaid": {}}))
     monkeypatch.setattr(qs_module.helpers, "get_imagemaid_root_path", lambda: "C:\\ImageMaid")
     monkeypatch.setattr(
-        qs_module,
+        workspace_status_module,
         "_probe_imagemaid_root_state",
         lambda _path: {"imagemaid_installed": True, "venv_python_exists": True},
     )
     monkeypatch.setattr(qs_module.database, "retrieve_section_data", lambda *_args, **_kwargs: None)
     monkeypatch.setattr(
-        qs_module,
+        workspace_status_module,
         "_validate_imagemaid_settings",
         lambda *_args, **_kwargs: (False, "missing_plex_validation", "Validate Plex first."),
     )
@@ -588,12 +588,12 @@ def test_workspace_app_readiness_imagemaid_missing_plex_points_to_plex(monkeypat
     assert payload["imagemaid"]["target_step"] == "010-plex"
 
 
-def test_workspace_app_readiness_kometa_freshness_keeps_app_available(monkeypatch, qs_module):
+def test_workspace_app_readiness_kometa_freshness_keeps_app_available(monkeypatch, qs_module, workspace_status_module):
     monkeypatch.setattr(qs_module.helpers, "get_menu_list", _template_list)
     monkeypatch.setattr(qs_module.database, "get_unique_config_names", lambda: ["cfg"])
-    monkeypatch.setattr(qs_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
+    monkeypatch.setattr(workspace_status_module, "_build_workspace_status_context", lambda *_args, **_kwargs: {"readiness": {}})
     monkeypatch.setattr(
-        qs_module,
+        workspace_status_module,
         "_build_final_gate",
         lambda *_args, **_kwargs: {
             "stage": "freshness",
@@ -601,12 +601,12 @@ def test_workspace_app_readiness_kometa_freshness_keeps_app_available(monkeypatc
             "todo_blockers": [],
         },
     )
-    monkeypatch.setattr(qs_module, "_build_kometa_install_context", lambda _name: {})
-    monkeypatch.setattr(qs_module, "_get_imagemaid_settings_section", lambda _name: ({}, {}))
+    monkeypatch.setattr(workspace_status_module, "_build_kometa_install_context", lambda _name: {})
+    monkeypatch.setattr(workspace_status_module, "_get_imagemaid_settings_section", lambda _name: ({}, {}))
     monkeypatch.setattr(qs_module.helpers, "get_imagemaid_root_path", lambda: "")
-    monkeypatch.setattr(qs_module, "_probe_imagemaid_root_state", lambda _path: {})
+    monkeypatch.setattr(workspace_status_module, "_probe_imagemaid_root_state", lambda _path: {})
     monkeypatch.setattr(qs_module.database, "retrieve_section_data", lambda *_args, **_kwargs: None)
-    monkeypatch.setattr(qs_module, "_validate_imagemaid_settings", lambda *_args, **_kwargs: (False, "", ""))
+    monkeypatch.setattr(workspace_status_module, "_validate_imagemaid_settings", lambda *_args, **_kwargs: (False, "", ""))
 
     payload = qs_module._build_workspace_app_readiness("cfg")
 
@@ -615,7 +615,7 @@ def test_workspace_app_readiness_kometa_freshness_keeps_app_available(monkeypatc
     assert payload["kometa"]["action_label"] == "Open Kometa"
 
 
-def test_workspace_status_context_aligns_app_steps_with_app_readiness(monkeypatch, qs_module):
+def test_workspace_status_context_aligns_app_steps_with_app_readiness(monkeypatch, qs_module, workspace_status_module):
     monkeypatch.setattr(qs_module.database, "retrieve_config_sections", lambda _name: [])
     monkeypatch.setattr(qs_module.database, "get_unique_config_names", lambda: ["cfg"])
     template_list = _template_list() + [("915-imagemaid.html", "ImageMaid")]
@@ -626,7 +626,7 @@ def test_workspace_status_context_aligns_app_steps_with_app_readiness(monkeypatc
             "imagemaid": {"state": "needs_setup"},
         }
 
-    monkeypatch.setattr(qs_module, "_build_workspace_app_readiness_from_status", fake_app_readiness)
+    monkeypatch.setattr(workspace_status_module, "_build_workspace_app_readiness_from_status", fake_app_readiness)
 
     ctx = qs_module._build_workspace_status_context("cfg", template_list, available_configs=["cfg"])
